+++ api/operator/v1beta1/vmauth_types_test.go
@@ -26,13 +26,13 @@ func TestVMAuthValidate(t *testing.T) {
 	// invalid ingress
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
 spec:
   ingress:
-    tlsHosts: 
+    tlsHosts:
       - host-1
       - host-2`,
 		wantErr: `spec.ingress.tlsSecretName cannot be empty with non-empty spec.ingress.tlsHosts`,
@@ -41,7 +41,7 @@ spec:
 	// both configSecret and external config is defined at the same time
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
@@ -57,13 +57,13 @@ spec:
 	// incorrect unauthorized access config, missing backends"
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
 spec:
   unauthorizedUserAccessSpec:
-    default_url: 
+    default_url:
       - http://url-1`,
 		wantErr: "incorrect cr.spec.UnauthorizedUserAccess syntax: at least one of `url_map`, `url_prefix` or `targetRefs` must be defined",
 	})
@@ -71,7 +71,7 @@ spec:
 	// incorrect unauthorized access config, bad metric_labels syntax
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
@@ -80,7 +80,7 @@ spec:
     metric_labels:
       124124asff: 12fsaf
     url_prefix: http://some-dst
-    default_url: 
+    default_url:
       - http://url-1`,
 		wantErr: `incorrect cr.spec.UnauthorizedUserAccess syntax: incorrect metricLabelName="124124asff", must match pattern="^[a-zA-Z_:.][a-zA-Z0-9_:.]*$"`,
 	})
@@ -88,7 +88,7 @@ spec:
 	// incorrect unauthorized access config url_map"
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
@@ -100,7 +100,7 @@ spec:
       - url_prefix: http://some-url
         src_paths: ["/path-1"]
       - url_prefix: http://some-url-2
-    default_url: 
+    default_url:
       - http://url-1`,
 		wantErr: `incorrect cr.spec.UnauthorizedUserAccess syntax: incorrect url_map at idx=1: incorrect url_map config at least of one src_paths,src_hosts,src_query_args or src_headers must be defined`,
 	})
@@ -108,12 +108,12 @@ spec:
 	// both unauthorizedUserAccessSpec and UnauthorizedUserAccess defined
 	f(opts{
 		src: `
-apiVersion: v1 
+apiVersion: v1
 kind: VMAuth
 metadata:
   name: must-fail
 spec:
-  unauthorizedAccessConfig: 
+  unauthorizedAccessConfig:
     - url_prefix: http://some-url
       src_paths: ["/path-1"]
     - url_prefix: http://some-url-2
@@ -124,7 +124,7 @@ spec:
     url_map:
       - url_prefix: http://some-url
         src_paths: ["/path-1"]
-    default_url: 
+    default_url:
       - http://url-1`,
 		wantErr: "at most one option can be used `spec.unauthorizedAccessConfig` or `spec.unauthorizedUserAccessSpec`, got both",
 	})