+++ api/operator/v1beta1/vmsingle_types.go
@@ -290,11 +290,17 @@ func (cr *VMSingle) PodLabels() map[string]string {
 
 // FinalLabels returns combination of selector and managed labels
 func (cr *VMSingle) FinalLabels() map[string]string {
-	v := cr.SelectorLabels()
+	return cr.ResourceLabels(nil)
+}
+
+// ResourceLabels returns combination of selector, resource and managed labels
+func (cr *VMSingle) ResourceLabels(input map[string]string) map[string]string {
+	var v map[string]string
 	if cr.Spec.ManagedMetadata != nil {
-		v = labels.Merge(cr.Spec.ManagedMetadata.Labels, v)
+		v = cr.Spec.ManagedMetadata.Labels
 	}
-	return v
+	v = labels.Merge(v, input)
+	return labels.Merge(v, cr.SelectorLabels())
 }
 
 // FinalAnnotations returns annotations to be applied for created objects