+++ internal/controller/operator/factory/vlsingle/vlsingle.go
@@ -48,8 +48,8 @@ func newPVC(r *vmv1.VLSingle) *corev1.PersistentVolumeClaim {
 		ObjectMeta: metav1.ObjectMeta{
 			Name:            r.PrefixedName(),
 			Namespace:       r.Namespace,
-			Labels:          labels.Merge(r.Spec.StorageMetadata.Labels, r.SelectorLabels()),
-			Annotations:     r.Spec.StorageMetadata.Annotations,
+			Labels:          r.ResourceLabels(r.Spec.StorageMetadata.Labels),
+			Annotations:     labels.Merge(r.FinalAnnotations(), r.Spec.StorageMetadata.Annotations),
 			OwnerReferences: []metav1.OwnerReference{r.AsOwner()},
 		},
 		Spec: *r.Spec.Storage,